Index.html no makes the apropriate ajax calls to get sunrise/sunset data from the test.py script.
test.py returns some rough HTML with with sunrise and sunset times in
UTC.  The date is currently hardcoded to 2016-05/25 (Today).

Still need to flush out the design of the json object sent to the server
side script and have the script return a proper json object and fix the
hard coded date in the server side script.

Also need to clean up my abhorent javascript but I'm still learning that
language.